// src/app/models/customer.model.ts
export interface Customer {
  _id?: string;
  name?: string;
  phone?: string;
  address?: string;
  lot?: string;
  block?: string;
  aptoBlock?: string;
  number?: number;
  apto?: string;
  district?: string;
  dateOfBith?: string;
  deliveryTax?: string;
  referencePoint?: string;
  comments?: string;
}